refactor(benchmarks): restructure into liander2024/ and custom/ with docs integration
- Move 3 Liander 2024 scripts to examples/benchmarks/liander2024/
- Move 6 custom benchmark scripts to examples/benchmarks/custom/
- Rename files for clarity (e.g. example_baseline -> custom_forecaster)
- Convert all scripts to jupytext format with markdown cells
- Hide boilerplate (SPDX, os.environ) with tags=[remove-cell]
- Add explanatory markdown sections between code cells
- Create README.md for each folder (overview, liander2024, custom)
- Integrate into docs build (myst-nb, excluded from execution)
- Add hierarchical sidebar nav (Benchmarking > Liander 2024 / Build Your Own)
- Group tutorials into sections (Getting Started, Model Training, Evaluation)
- Extend CI tasks (notebooks, notebooks-check, notebooks-clear)
- Update tools/check_notebook_outputs.py for new paths
- Add per-file-ignores for benchmark lint rules
