+++ examples/benchmarks/liander2024/__init__.py
@@ -1,0 +1,4 @@
+# SPDX-FileCopyrightText: 2025 Contributors to the OpenSTEF project <openstef@lfenergy.org>
+#
+# SPDX-License-Identifier: MPL-2.0
+"""Liander 2024 benchmark scripts."""
