+++ examples/benchmarks/liander2024/compare_benchmark_runs.py
@@ -1,8 +1,45 @@
-"""Example for comparing benchmark results from different runs on the Liander 2024 dataset."""
+# ---
+# jupyter:
+#   jupytext:
+#     formats: ipynb,py:percent
+#     text_representation:
+#       extension: .py
+#       format_name: percent
+#       format_version: '1.3'
+#       jupytext_version: 1.19.1
+#   kernelspec:
+#     display_name: Python 3
+#     language: python
+#     name: python3
+# ---
+
+# %% [markdown]
+# # Compare Benchmark Runs
+#
+# Generate side-by-side comparison plots from multiple benchmark runs on the
+# Liander 2024 dataset.
+#
+# **Prerequisites:** Run at least two models first (e.g. XGBoost + GBLinear via
+# the *XGBoost & GBLinear* notebook).
+#
+# **What this does:**
+#
+# 1. Loads results from multiple model runs (each stored in its own directory)
+# 2. Computes metrics across all targets using
+#    [`BenchmarkComparisonPipeline`](https://openstef.github.io/openstef/v4/api/generated/openstef_beam.benchmarking.BenchmarkComparisonPipeline.html)
+# 3. Produces comparison visualizations (boxplots, ranking tables, per-target breakdowns)
+
+# %% tags=["remove-cell"]
 # SPDX-FileCopyrightText: 2025 Contributors to the OpenSTEF project <openstef@lfenergy.org>
 #
 # SPDX-License-Identifier: MPL-2.0
 
+# %% [markdown]
+# ## Setup
+#
+# Point at the result directories from your benchmark runs.
+
+# %%
 from pathlib import Path
 
 from openstef_beam.analysis.models import RunName
@@ -18,6 +55,13 @@ OUTPUT_PATH = BASE_DIR / "./benchmark_results_comparison"
 BENCHMARK_DIR_GBLINEAR = BASE_DIR / "benchmark_results" / "GBLinear"
 BENCHMARK_DIR_XGBOOST = BASE_DIR / "benchmark_results" / "XGBoost"
 
+# %% [markdown]
+# ## Load run results
+#
+# Each run is identified by a name and backed by a `LocalBenchmarkStorage` that
+# points at the directory where that model's results were saved.
+
+# %%
 check_dirs = [
     BENCHMARK_DIR_GBLINEAR,
     BENCHMARK_DIR_XGBOOST,
@@ -32,6 +76,13 @@ run_storages: dict[RunName, BenchmarkStorage] = {
     "xgboost": LocalBenchmarkStorage(base_path=BENCHMARK_DIR_XGBOOST),
 }
 
+# %% [markdown]
+# ## Run comparison
+#
+# The pipeline loads predictions from each run, re-evaluates them with the
+# Liander 2024 analysis config, and produces comparison visualizations.
+
+# %%
 target_provider = create_liander2024_benchmark_runner(
     storage=LocalBenchmarkStorage(base_path=OUTPUT_PATH),
 ).target_provider
